toolkit: Port AccessibleStatusBarTest to test::AccessibleTestBase
Now we have test::AccessibleTestBase we can use it instead of
performing some setup operations manually.

The only actual change is that we don't pass OpenFlags=S when creating
an Impress document, but it never seemed useful (it was only ported
from Java out of zeal), and I can't even find what the S option is for.
In all cases, this doesn't affect the tests themselves.

Change-Id: Ic8d56dd9bc3ec957891868b23bb241ba3bf33049
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147003
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>