tdf#153520 vcl: Align listbox invalidation with mouseover check
Before

    commit b7afca98f5107689190ee5760bf183203af76dc9
    Date:   Fri Oct 7 16:40:23 2022 +0100

        Related: tdf#98067 do RollOver for Edit as well as SpinButton

, the check whether the mouse is over a control
was only checking the control's child windows.
The above commit introduced a check for the control's
window as well in `ImplSmallBorderWindowView::DrawWindow`.

When moving the mouse inside or outside a `ListBox` control,
the invalidation of the corresponding border window happened
when the mouse cursor would enter or leave the `ListBox`'s
`ImplWin` child (in `ImplWin::PreNotify`).

The `ImplWin` only covers a part of the `ListBox` area. As
a consequence, when moving the mouse out of the `ImplWin`
area, the mouse would still be over the parent `ListBox` area,
so the `ControlState::ROLLOVER` state would still be set
in `ImplSmallBorderWindowView::DrawWindow` from the above
commit on.

Since there was no additional invalidation when the mouse
moved out of the `ListBox` area, the `ROLLOVER` state would
remain set even when the mouse left the `ListBox` area as well.

Move the invalidation from `ImplWin` to `ListBox` so
this gets processed when the mouse cursor actually enters or
leaves the `ListBox`, not just the `ImplWin` child.

Also align the mouse over check in
`ImplWin::ImplDraw` with the one in
`ImplSmallBorderWindowView::DrawWindow` again.
(Take the parent into account there as well.)

Change-Id: I8974e370819719489e3d9f091d96b3353888d26b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147018
Tested-by: Jenkins
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>