tdf#153614 qt: Set keyboard focus state when focused, too
For example the Breeze theme only draws a light blue
focus rectangle for list boxes when the
`QStyle::State_KeyboardFocusChange` state is set in addition
to `QStyle::State_HasFocus`.
Therefore, set that state in addition to ensure that the focused
control actually gets a focus indicator when moving there
using the keyboard.

Change-Id: Ib4b85f9140629e6b69c80b85e85913392a6c000f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147019
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>