tdf#153613 writerfilter: don't always split para for shape w/ pagebreak
The problem was that an extra paragraph was added,
and the shape was anchored to the wrong page.

There are lots of places that handle deferred breaks,
and they typically do this only for FirstParagraph or non-first
character properties run.

This is needed here too.
make CppunitTest_sw_ooxmlexport18 CPPUNIT_TEST_NAME=testTdf153613_textboxAfterPgBreak
make CppunitTest_sw_ooxmlexport18 CPPUNIT_TEST_NAME=testTdf153613_textboxAfterPgBreak2

Of course, RTF is goofy, so handle that as an exception
to just leave it alone as it has been.
To solve it properly is likely very tricky - as are all things RTF,
because we can't trust IsFirstRun in general and especially not for RTF.
make CppunitTest_sw_rtfexport5 CPPUNIT_TEST_NAME=testN818997B

Change-Id: Iad06f3ea9684a2860ae6ba724887bdf0b5850f94
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147169
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>