sw floattable: limit vertical position for the follows of split flys
The bugdoc has 2 pages and the 2nd page had a fly frame with an
unexpected vertical shift towards the bottom.

The problem is that the floattable has a vertical offset defined, but
Word only handles that on the first page of the split table, not on
later pages. This makes sense: we split the table to multiple pages so
that the content fits, an additional offset on follow pages would result
in a lot of additional pages.

Fix the problem by adjusting
SwToContentAnchoredObjectPosition::CalcPosition(), so that the vertical
offset is only applied on master flys, not on follow flys.

This implicitly means that the behavior is unchanged for non-split flys,
since those always have a first page, which is nicely consistent.

Change-Id: I8c86cd2df6868cc5c1a10dfec359f3e02f5c1102
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147141
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins