tdf#153613 DOCX: DID YOU FIX ME? unit test
So once we have a page break, we need to know the rest
of the paragraph's contents before the decision can
be made to split the paragraph or not.

In order to do this properly, the shape processing would need
to be delayed until some "normal" character run is seen
before a carriage return.

That's too complicated for me,
and probably not worth doing anyway.

make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf153613_anchoredAfterPgBreak6

Change-Id: Ie79574f2e1656a230da8df1b95fe61ea96c2802a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147175
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>