tdf#153105 PPTX export: fix "Custom position/size" background image
Map size and the 9 preset positions of the ODF background image
style "Custom position/size" to the OOXML a:stretch/a:fillRect
with the appropriate left/top/right/bottom arguments.

Note: it seems, applying a:stretch or a:tile was not mandatory,
but missing a:stretch resulted non-editable document in Office 365.

Note: the import of the PPTX mapping hasn't been implemented, yet.

Follow-up to commit 224c58e35b1874c8aa031295021bc212fa71bdff
"tdf#153107 OOXML export: fix scale of tile of shape background".

Change-Id: Ie940ebc2610c5b75126da05678a04ed1552cacb3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147337
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>