tdf#142942 Drop unused checkbox col in sw AutoCorrect opts dialog
While the extra checkbox column enabled by

    weld::TreeView::enable_toggle_buttons(weld::ColumnToggleType::Check)

is used for the non-Writer case in the "Options" and
"Localized Options" tabs in the AutoCorrect options
dialog, they are not used in the case of Writer,
where there are two columns of checkboxes handled
separately ("[M]: Replace while modifying existing text"
and "[T]: AutoCorrect while typing").

Before

    commit c0aef006c3c2df01dd5cb746df56a81adbe3485a
    Date:   Tue Jun 30 15:03:31 2020 +0100

        tdf#134270 rearrange so we know the final treeview width

, setting column widths manually by calling
`weld::TreeView::set_column_fixed_widths` would disable
the extra space for the (unused) checkboxes (by setting
`m_bDisableCheckBoxAutoWidth`).

Since that commit, it no longer does, causing the
separately handled real checkbox in the first visible
column to be unnecessarily positioned further to the right
in this dialog for Writer for the non-gtk case.

Fix this by dropping the use of the unused
checkbox column for the Writer case, i.e. the `bShowSWOptions`
case in `OfaQuoteTabPage::OfaQuoteTabPage` and
in `OfaSwAutoFmtOptionsPage::OfaSwAutoFmtOptionsPage`.
The handling for the non-Writer modules (i.e. `!bShowSWOptions`
case in `OfaQuoteTabPage::OfaQuoteTabPage` and
`OfaAutocorrOptionsPage`) remains unchanged.

Change-Id: If7bd216fc4efda506817cffb55230722bb2ac89b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147454
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>