tdf#153526 writerfilter: catch exception from no NumberingRules
This fixes a LO 7.5 regression (well, exposed some other flaws) from
commit 9a0c1c9451543c82c578a96d549ec5ea5233d07f.

I first tried testing if hasPropertyByName,
but that returns true(????), even though this is an editeng component.

Well, ignoring the fundamental issues at play here
(and there are many - like having a comment
take over m_xPreviousParagraph)
the issue is easily solved with a try/catch
instead of a redundant if clause.

Change-Id: I4f27fce3e2984235df19dc3ed4be558891b28a90
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147486
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>