tdf#151548 sw content controls: add UI for Id and TabIndex
This is my first go at adding anything to the UI.
(I've done some fixing of other people's design,
but never created anything new myself.)

I took all my inspiration from Miklos'
d8d4e2e2de8911e476d71d72124dc6356c9d6c5d.

In the first iteration I did a fancy "same as id"
scenario. But ID-as-unsigned, securing unique IDs,
and English-only number recognition
sadly made me decide to drop the fancy features.

Now, both are simple, stand-alone spinbuttons
that depend on tooltips to guide user input, not fancy comboboxes.

It is important for the ID to display as unsigned,
since that is how it is used as a "name" in VBA.
It should be copy-pasteable for use in VBA programming.

It is useful for the TABINDEX to display as signed,
since it allows entering a -1 to disable tab navigation to the control.

make UITest_sw_ui_misc

Change-Id: Ia7c99888e60c33ac39616b24c7c1715604c3ec69
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147591
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>