word vba: tie "Word" to VBAGlobals
This allows Word. to prefix global commands.
While this is superfluous to use this profix in a DOC/X,
the purpose is to allow the spreadsheet to call
Word.<whatever> macro calls that execute in a word processor.

At this point, it is simply added to SwVbaGlobals
to allow macro calls to run without raising an error.

make CppunitTest_sw_macros_test

Change-Id: Ibeda67d458d4757be809d8669a4e424c4e6c3458
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147682
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>