Hide Quickstarter from General options in Windows store apps
In that environment, the shell:Startup shortcut created by us points to the
quickstart.exe under Program Files\WindowsApps\<StoreSpecificAppDir>\program.
The resulting shortcut can't launch, because of an inaccessible part in the
path (the protected WindowsApps directory). So the feature does not work in
the dialog.

Windows store apps should advertise the Windows startup task in the manifest.
Let's just hide the dialog entry in this case. A possible TODO would be to
provide a button instead, which would open the Startup Apps system applet,
similar to what we do for Default apps for file associations.

Change-Id: Ieafdf6d23ced96506b01c6b3cf9fb12904696df1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147735
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>