tdf#114441 Convert sal_uLong to better data type sal_uInt32
In this patch, all the instances of sal_uLong in sw/source/ui/fldui/
folder are changed to better types.

The variable m_nOldFormat is changed to the type sal_uInt32 in fldfunc.hxx
as we have m_nOldFormat = GetCurField()->GetFormat() in fldfunc.cxx file
and GetFormat() return sal_uInt32. Also, the data type of variable nFormat
is changed from to sal_uInt32 as there is a conditional operation with the
sal_uInt32 output. This is also the case for other changed files.

The variable nFormat is changed to sal_uInt32 in files fldref.cxx and
fldvar.cxx.

Change-Id: I57ac03e48d4f0eecc96cfdbc19e92e00148f07d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147881
Tested-by: Hossein <hossein@libreoffice.org>
Reviewed-by: Hossein <hossein@libreoffice.org>