tdf#114287 xmloff: ODF import: fix text:list override of list style
The problem is that the list element applies a list style, which should
override the list style that's set in the common style
"myStyleWithNumbering" of the paragraph, but these happend to be the
same list styles and so the override was skipped, resulting in bad text
formatting.

<text:list xml:id="list3425910577850" text:continue-numbering="true" text:style-name="Numbering_20_123">

The same problem exists if the list-item element has a style-override.

This is similar to bug i#101349 which fixed it for an automatic
paragraph style applying the same list style as its parent style.

Change-Id: Idc6f10f93976de7e7bdceccbe3012c4157f61ad1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147930
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>