tdf#131386 ApplyParagraphMarkFormatToNumbering remove RTF fallback
This stops run-time evaluation of the last char-run properties
on the paragraph - which were then applied to the numbering
character itself. This was done to imitate MS Word's ability
to format the CR itself - which similarly affected the numbering.

This awful hack has been slowly evicted as RES_PARATR_LIST_AUTOFMT
was created and rolled out to DOCX in 6.4, DOC in 7.5,
and now RTF in 7.5.1.

A change made in 7.2 (also made temporarily in 6.4 development)
looked not only at the start == end char run,
but also any run going past it/up to the end.

Although the 7.2 change helped some documents, it hurt others.
It can now be eliminated completely since RES_PARATR_LIST_AUTOFMT
handles that situation at import time.

Since this is a run-time layout adjustment, removing it should
have no impact on export. Any effect made to MS Format -> ODT
is likely incorrect as seen by various bug reports
like tdf#146168 and tdf#150613.

In the worst case, it would only be a few recent (7.2 - 7.5)
DOC -> ODT or RTF -> ODT migrations that will "look different".
Chalk those up to implementation errors that need to be corrected.

Change-Id: I4ae76008ffcbc5f19580b6ddca08197b7074e12c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148014
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>