sw floattable: update layout when enabling fly split via UNO
Changing e.g. the HoriOrientPosition property resulted in a layout
update, but setting IsSplitAllowed to true had no effect on the layout.

The problem seems to be that the new RES_FLY_SPLIT is not handled in any
update notification.

Fix the problem by initially handling this the same way as RES_FRM_SIZE
is already done, that triggers a 2nd page when a large fly frame is
newly allowed to split, for existing fly frames as well.

Disabling needs more work, still.

Change-Id: I2cba04213286e72384a56a28f2b321f7fd577639
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148029
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins