sw floattable: fix vertical position of follow flys with rel orient = page
The problem was that the follow frame's top was the 2nd page's top,
because we ignore the vertical offset for follow fly frames.

Ignoring the original offset is fine, but follow frames are still inside
the body frame in Word, so their offset is effectively the header / top
margin area.

Fix the problem by improving how nRelPosY is tweaked for follow fly
frames: an additional positive offset is wanted when the rel orient is
page.

With this, this bugdoc now renders perfectly.

Change-Id: Iec4b85339a7b4b69944d25eb2e06a6e3f977f093
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148086
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins