Resolves: tdf#153885 keep popovers with application window parent inside that
for wayland.  If the popover window is the application window, contrain
it within the application window so it won't be cut off off screen,
because gtk under wayland just puts it when it wants to be regardless of
screen bounds.

Leave dialog hosted ones alone, like format, watermark, which are likely
presented in the middle of the screen and are too small to constrain the
popover inside.

Change-Id: Ibdc0749613a3b587414e88be7d0aea81c637f31a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148091
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>