tdf#152073 tdf#153895 basicide: Set bg color for border win
As described in the original commit to address tdf#152073
("Glitch rendering line numbers in Basic IDE editor (kf5 only)"),
the fact that the KDE Breeze style's frame border is partly
transparent was causing artifacts in the Basic IDE:

    commit bf1b3c526a20b7a1f7bd19f8a29e9dc6a915646d
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jan 31 16:08:53 2023 +0100

        tdf#152073 qt: Draw background when drawing frame

        For native drawing of a frame (border), use
        the window background color as default color.

        As mentioned in commit 1a07b1ade15cac66043119ecbd6e97673d9edc21
        ("tdf#138010 (IV) VclScrolledWindow: Use actual border width"),
        the Qt/KDE Breeze style uses a frame width of 2, with the
        actual 1 pixel border being surrounded by a 1 pixel
        padding/margin.

        Transparent background did not ensure that the 1 pixel
        padding/margin is repainted and could thus result
        in artifacts from what was painted to that location
        previously (s. the Basic IDE editor example from
        tdf#152073).

        [...]

However, there are cases when a transparent background
is actually wanted, like in the print dialog (tdf#153895).

Therefore, revert the original commit mentioned above
and explicitly set a background color for the border windows
of the involved windows in the Basic IDE instead, which is
in line with the alternative already mentioned in the above commit
("An alternative might be to explicitly draw the
background further up the call stack, [...]."), but enables
transparency in the qt5/qt6/kf5-specific drawing for
the frame again.

Change-Id: Ifb4deb80c6ae245ec630a4fdc23f675e5587836c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/148117
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>